Relocate `/profile shikimori` to `/shikimori profile`
